Update card styles and course information
This commit updates the card styles in the `_card.scss` file to include an overflow property and sets the image inside the card to have a cover object-fit. It also updates the course information in the `courses.js` file to fix a typo and change the image URL. Additionally, it updates the content in the `dev.js` file to include a description of static site generation and update the website links. Finally, it adds a new image file `ai_webinar.jpg`.