Update styles and links on dev page
This commit updates the styles and links on the dev page. It adds a max-width to the dev-link images and hides the sidebar on small screens. It also updates the links to mikkelbech.com and blackboxpedals.com. Additionally, it adds a new section for tools and includes a link to the Additive Prompt Builder.